// src/generated/models/Cr871_appusersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr871_appuserscr871_role = {
  144610000: 'CISO',
  144610001: 'Assessor'
} as const;
export type Cr871_appuserscr871_role = keyof typeof Cr871_appuserscr871_role;
export const Cr871_appusersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr871_appusersstatecode = keyof typeof Cr871_appusersstatecode;
export const Cr871_appusersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr871_appusersstatuscode = keyof typeof Cr871_appusersstatuscode;

export interface Cr871_appusersBase {
  cr871_appuserid: string;
  cr871_email?: string;
  cr871_name?: string;
  cr871_role?: Cr871_appuserscr871_role;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr871_appusersstatecode;
  statuscode?: Cr871_appusersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr871_appusers extends Cr871_appusersBase {
  cr871_rolename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}